+++ src/CSG.h
@@ -13,20 +13,26 @@ CSGInterval getInterval(const Hittable& obj, const Ray& r);
 class CSGUnion : public Hittable {
     std::shared_ptr<Hittable> left, right;
 public:
-    CSGUnion(std::shared_ptr<Hittable> l, std::shared_ptr<Hittable> r): left(std::move(l)), right(std::move(r)) {}
+    CSGUnion(std::shared_ptr<Hittable> l, std::shared_ptr<Hittable> r)
+        : left(std::move(l)), right(std::move(r)) {
+    }
     bool hit(const Ray& r, Interval rayT, HitRecord& rec) const override;
 };
 
 class CSGIntersection : public Hittable {
     std::shared_ptr<Hittable> left, right;
 public:
-    CSGIntersection(std::shared_ptr<Hittable> l, std::shared_ptr<Hittable> r): left(std::move(l)), right(std::move(r)) {}
+    CSGIntersection(std::shared_ptr<Hittable> l, std::shared_ptr<Hittable> r)
+        : left(std::move(l)), right(std::move(r)) {
+    }
     bool hit(const Ray& r, Interval rayT, HitRecord& rec) const override;
 };
 
 class CSGDifference : public Hittable {
     std::shared_ptr<Hittable> left, right;
 public:
-    CSGDifference(std::shared_ptr<Hittable> l, std::shared_ptr<Hittable> r): left(std::move(l)), right(std::move(r)) {}
+    CSGDifference(std::shared_ptr<Hittable> l, std::shared_ptr<Hittable> r)
+        : left(std::move(l)), right(std::move(r)) {
+    }
     bool hit(const Ray& r, Interval rayT, HitRecord& rec) const override;
 };